requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
spacy
neotime
neo4j
psycopg2-binary
SQLAlchemy
elasticsearch
opensearch-py
numpy
pandas
scikit-learn
mlflow
pydantic
python-dotenv
pyyaml
langdetect
syntok
celery
redis
ruff
black
mypy
pytest
fpdf2
